requirements.txt
# FingerPrinter
# A tool to generate file fingerprints
# Github://www.github.com/0x4248/FingerPrinter
# Licence: GNU General Public License v3.0
# By: 0x4248

PyYAML==6.0.1